refactor(game)!: remove async scene update loop
It's next to impossible to implement this correctly (and it wasn't), as
there is a *lot* of shared mutable state.
A quick run with thread sanitizer will show you how broken it is.

Expect up to 10% FPS decrease while playing for normal builds.
Debug build are impacted much more but that can be improved later.

Next stop is removing async loop from GenericInfoUpdater and using an
input queue instead of callbacks.