refactor(game): drop separate handling of csv lines with backslashes
This can not possibly work in my understanding.
Skins use backslashes in file paths, so it's impossible to properly
understand whether it's a path separator or an escape.
We could consider `\,` as escapes, always, but it doesn't seem like
that works in LR2 and I can't find skins that depend on this.